+++ payment-method-messaging/src/main/java/com/stripe/android/paymentmethodmessaging/element/PaymentMethodMessagingElement.kt
@@ -3,7 +3,6 @@ package com.stripe.android.paymentmethodmessaging.element
 import android.app.Application
 import androidx.annotation.ColorInt
 import androidx.annotation.FontRes
-import androidx.annotation.RestrictTo
 import androidx.compose.material.MaterialTheme
 import androidx.compose.runtime.Composable
 import androidx.compose.runtime.getValue
@@ -16,7 +15,6 @@ import java.util.Locale
 import javax.inject.Inject
 
 @PaymentMethodMessagingElementPreview
-@RestrictTo(RestrictTo.Scope.LIBRARY_GROUP)
 class PaymentMethodMessagingElement @Inject internal constructor(
     private val messagingCoordinator: PaymentMethodMessagingCoordinator,
     private val eventReporter: PaymentMethodMessagingEventReporter
@@ -47,7 +45,6 @@ class PaymentMethodMessagingElement @Inject internal constructor(
         content?.Content(appearanceState)
     }
 
-    @RestrictTo(RestrictTo.Scope.LIBRARY_GROUP)
     companion object {
         fun create(application: Application): PaymentMethodMessagingElement {
             return DaggerPaymentMethodMessagingComponent.builder()
@@ -59,33 +56,28 @@ class PaymentMethodMessagingElement @Inject internal constructor(
     /**
      * The result of a [configure] call.
      */
-    @RestrictTo(RestrictTo.Scope.LIBRARY_GROUP)
     sealed interface ConfigureResult {
 
         /**
          * The configuration succeeded and [Content] will display a view.
          */
-        @RestrictTo(RestrictTo.Scope.LIBRARY_GROUP)
         class Succeeded internal constructor() : ConfigureResult
 
         /**
          * The configuration succeeded but no content is available to display. (e.g. the amount is less than the
          * minimum for available payment methods).
          */
-        @RestrictTo(RestrictTo.Scope.LIBRARY_GROUP)
         class NoContent internal constructor() : ConfigureResult
 
         /**
          * The configure call failed e.g. due to network failure or because of an invalid [Configuration].
          */
-        @RestrictTo(RestrictTo.Scope.LIBRARY_GROUP)
         class Failed internal constructor(val error: Throwable) : ConfigureResult
     }
 
     /**
      * Configuration for [PaymentMethodMessagingElement].
      */
-    @RestrictTo(RestrictTo.Scope.LIBRARY_GROUP)
     class Configuration {
         private var amount: Long? = null
         private var currency: String? = null
@@ -158,7 +150,6 @@ class PaymentMethodMessagingElement @Inject internal constructor(
         }
     }
 
-    @RestrictTo(RestrictTo.Scope.LIBRARY_GROUP)
     class Appearance {
         private var theme: Theme = Theme.LIGHT
         private var font: Font.State? = null
@@ -202,14 +193,12 @@ class PaymentMethodMessagingElement @Inject internal constructor(
         /**
          * The theme of the payment method icons to display.
          */
-        @RestrictTo(RestrictTo.Scope.LIBRARY_GROUP)
         enum class Theme {
             LIGHT,
             DARK,
             FLAT
         }
 
-        @RestrictTo(RestrictTo.Scope.LIBRARY_GROUP)
         class Font {
             private var fontFamily: Int? = null
             private var fontSizeSp: Float? = null
@@ -260,7 +249,6 @@ class PaymentMethodMessagingElement @Inject internal constructor(
             )
         }
 
-        @RestrictTo(RestrictTo.Scope.LIBRARY_GROUP)
         class Colors {
             private var textColor: Int = StripeThemeDefaults.colorsLight.onComponent.toArgb()
             private var infoIconColor: Int = StripeThemeDefaults.colorsLight.subtitle.toArgb()