+++ payments-core/src/main/java/com/stripe/android/challenge/confirmation/IntentConfirmationChallengeArgs.kt
@@ -7,5 +7,6 @@ import kotlinx.parcelize.Parcelize
 @Parcelize
 internal data class IntentConfirmationChallengeArgs(
     val publishableKey: String,
+    val productUsage: List<String>,
     val intent: StripeIntent
 ) : Parcelable