+++ paymentsheet/src/main/java/com/stripe/android/paymentelement/EmbeddedPaymentElement.kt
@@ -510,9 +510,7 @@ class EmbeddedPaymentElement @Inject internal constructor(
              * By default, the embedded payment element offers a card scan button within the new card entry form.
              * When opensCardScannerAutomatically is set to true,
              * the card entry form will initialize with the card scanner already open.
-             * **Note**: The stripecardscan dependency must be added to set `opensCardScannerAutomatically` to true
              */
-            @RestrictTo(RestrictTo.Scope.LIBRARY_GROUP)
             fun opensCardScannerAutomatically(opensCardScannerAutomatically: Boolean) = apply {
                 this.opensCardScannerAutomatically = opensCardScannerAutomatically
             }