+++ paymentsheet/src/main/java/com/stripe/android/paymentelement/confirmation/PaymentMethodConfirmationOption.kt
@@ -1,5 +1,6 @@
 package com.stripe.android.paymentelement.confirmation
 
+import com.stripe.android.model.PaymentMethod
 import com.stripe.android.model.PaymentMethodCreateParams
 import com.stripe.android.model.PaymentMethodExtraParams
 import com.stripe.android.model.PaymentMethodOptionsParams
@@ -9,8 +10,6 @@ import com.stripe.android.paymentsheet.PaymentSheet
 import kotlinx.parcelize.Parcelize
 
 internal sealed interface PaymentMethodConfirmationOption : ConfirmationHandler.Option {
-    val passiveChallengeComplete: Boolean
-    val attestationComplete: Boolean
     val optionsParams: PaymentMethodOptionsParams?
 
     fun updatedForDeferredIntent(
@@ -26,8 +25,6 @@ internal sealed interface PaymentMethodConfirmationOption : ConfirmationHandler.
         val originatedFromWallet: Boolean = false,
         val hCaptchaToken: String? = null,
         val attestationToken: String? = null,
-        override val passiveChallengeComplete: Boolean = false,
-        override val attestationComplete: Boolean = false,
     ) : PaymentMethodConfirmationOption {
         override fun updatedForDeferredIntent(
             intentConfiguration: PaymentSheet.IntentConfiguration,
@@ -48,8 +45,8 @@ internal sealed interface PaymentMethodConfirmationOption : ConfirmationHandler.
         override val optionsParams: PaymentMethodOptionsParams?,
         val extraParams: PaymentMethodExtraParams?,
         val shouldSave: Boolean,
-        override val passiveChallengeComplete: Boolean = false,
-        override val attestationComplete: Boolean = false,
+        val passiveChallengeComplete: Boolean = false,
+        val attestationComplete: Boolean = false,
     ) : PaymentMethodConfirmationOption {
 
         override fun updatedForDeferredIntent(