+++ paymentsheet/src/main/java/com/stripe/android/paymentelement/confirmation/intent/CustomerSheetIntentConfirmationModule.kt
@@ -1,0 +1,17 @@
+package com.stripe.android.paymentelement.confirmation.intent
+
+import dagger.Binds
+import dagger.Module
+
+@Module(includes = [IntentConfirmationModule::class])
+internal interface CustomerSheetIntentConfirmationModule {
+    @Binds
+    fun bindsSetupIntentInterceptorFactory(
+        defaultCustomerSheetSetupIntentInterceptorFactory: DefaultCustomerSheetSetupIntentInterceptorFactory
+    ): CustomerSheetSetupIntentInterceptor.Factory
+
+    @Binds
+    fun bindsIntentConfirmationInterceptorFactory(
+        customerSheetConfirmationInterceptorFactory: CustomerSheetIntentConfirmationInterceptorFactory
+    ): IntentConfirmationInterceptor.Factory
+}