+++ CHANGELOG.md
@@ -2,8 +2,12 @@
 
 ## XX.XX.XX - 20XX-XX-XX
 
+### Payment Method Messaging
+* [Added] Payment Method Messaging Element is now available in public preview.
+
 ### PaymentSheet
 * [Added][12011](https://github.com/stripe/stripe-android/pull/12011) Builder APIs for PaymentSheet appearance customization (Colors, Shapes, Typography, PrimaryButtonColors, and embedded row styles)
+* [Added] opensCardScannerAutomatically configuration is now available.
 
 ### Payments
 * [Added][12014](https://github.com/stripe/stripe-android/pull/12014) `ConfirmationToken.PaymentMethodPreview` now includes structured `Card` field with detailed card information (brand, country, expiry, funding, last4, etc.)